requirements.txt
matplotlib==3.*
numpy==1.*
scipy==1.*
tensorflow-cpu==2.*
